refactored some code, divided funcs into modules
