+++ js/waiterWithShift.js
@@ -1,0 +1,3 @@
+//after executing the following function, we know that we have a waiter that is
+//linked to a shift collection
+module.exports =
