+++ views/registration.handlebars
@@ -8,7 +8,7 @@
     <input class="buttonStyle" id="regPageRegButton" type="submit" name="submitRegistration" value="Register">
   </form>
 
-  <p>{{output}}</p>
+  <p id="regComplete">{{output}}</p>
 
   <p><a href="/">Back to Login Page</a></p>
 </div>
